Use dtype= instead of deprecated torch_dtype=

+++ run.py
@@ -63,7 +63,7 @@ def main():
     if tokenizer.pad_token is None:
         tokenizer.pad_token = tokenizer.eos_token
     model = AutoModelForCausalLM.from_pretrained(
-        args.model, torch_dtype=dtype, device_map="auto",
+        args.model, dtype=dtype, device_map="auto",
         trust_remote_code=True,
     )
     model.eval()
